+++ spec/helpers/reporter.ts
@@ -1,0 +1,6 @@
+import {DisplayProcessor, SpecReporter} from 'jasmine-spec-reporter';
+
+jasmine.getEnv().clearReporters();
+jasmine.getEnv().addReporter(new SpecReporter({
+    customProcessors: [DisplayProcessor]
+}));