+++ CHANGELOG.md
@@ -5,7 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+[Unreleased]
+------------
+### Added
+- Added a visual spec reporter for [Jasmine](https://jasmine.github.io/).
+
 0.1.0 - 2018-04-07
 ------------------
 ### Added
 - Initial release.
+
+[Unreleased]: https://github.com/jbenner-radham/typescript-mothership-demo/compare/v0.1.0...HEAD